pyproject.toml
[project]
name = "bd-intelligence"
version = "0.1.0"
description = "Pharma BD decision intelligence tool using LangGraph multi-agent system"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.88",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "langchain-anthropic>=1.4.0",
    "langchain-aws>=1.4.3",
    "langgraph>=1.1.3",
    "mcp>=1.27.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
    "uvicorn>=0.44.0",
]

[tool.uv]
python = "3.11"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
